refactor: break out utils for migration tests into test util.
